+++ src/fastmath/interpolation/kriging.clj
@@ -13,7 +13,7 @@
                 (for [xs2 xss]
                   (variogram (distance xs1 xs2))))
               (mat/rows->RealMatrix))]
-    (if (or (and (number? error) (m/zero? error)) (not (seq error)))
+    (if (or (and (number? error) (m/zero? (double error))) (not (seq error)))
       m
       (mat/sub m (MatrixUtils/createRealDiagonalMatrix
                   (m/seq->double-array (if (sequential? error)