+++ src/fastmath/interpolation/shepard.clj
@@ -7,13 +7,14 @@
 #_(set! *warn-on-reflection* true)
 
 (defn shepard
-  [xss ys {:keys [^double p distance]
-           :or {p 2.0}}]
-  (let [dist (or distance (if (number? (first xss)) d/euclidean-1d d/euclidean))
-        cache (into {} (map vector xss ys))
-        p- (m/- p)]
-    (fn ^double [xs]
-      (or (cache xs)
-          (->> (map (fn [xs']
-                      (m/pow (dist xs xs') p-)) xss)
-               (stats/wmean ys))))))
+  ([xss ys] (shepard xss ys))
+  ([xss ys {:keys [^double p distance]
+            :or {p 2.0}}]
+   (let [dist (or distance (if (number? (first xss)) d/euclidean-1d d/euclidean))
+         cache (into {} (map vector xss ys))
+         p- (m/- p)]
+     (fn ^double [xs]
+       (or (cache xs)
+           (->> (map (fn [xs']
+                       (m/pow (dist xs xs') p-)) xss)
+                (stats/wmean ys)))))))